requirements.txt
Flask
requests
firebase-admin
python-dotenv
gunicorn
Flask-Cors
python-dotenv
